pyproject.toml
[project]
name = "wizelit-mcp"
version = "0.1.0"
description = "Minimal Wizelit project with MCP servers and SDK integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.3",
    "boto3>=1.42.5",
    "langchain>=1.1.3",
    "langgraph>=1.0.4",
    "langchain-aws>=1.1.0",
    "mcp>=1.23.3",
    "PyGithub>=2.4.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "crewai>=1.6.1",
    "redis>=4.5.0",
    "uvicorn>=0.30.0",
    "wizelit-sdk @ git+ssh://git@github.com/duongle-wizeline/wizelit-sdk.git@v0.1.21"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["code_scout*", "mcp_servers*", "utils*"]
exclude = ["public*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
